+++ ydb/core/transfer/ut/common/utils.h
@@ -68,6 +68,26 @@ inline bool Checker<bool>::Get(const ::Ydb::Value& value) {
 }
 
 template<>
+inline i8 Checker<i8>::Get(const ::Ydb::Value& value) {
+    return value.int32_value();
+}
+
+template<>
+inline i16 Checker<i16>::Get(const ::Ydb::Value& value) {
+    return value.int32_value();
+}
+
+template<>
+inline i32 Checker<i32>::Get(const ::Ydb::Value& value) {
+    return value.int32_value();
+}
+
+template<>
+inline i64 Checker<i64>::Get(const ::Ydb::Value& value) {
+    return value.int64_value();
+}
+
+template<>
 inline ui32 Checker<ui32>::Get(const ::Ydb::Value& value) {
     return value.uint32_value();
 }