+++ ydb/core/tx/conveyor_composite/service/process.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "common.h"
+#include "scope.h"
 #include "worker.h"
 
 #include <ydb/core/tx/conveyor_composite/usage/config.h>
@@ -8,6 +9,7 @@
 #include <ydb/library/accessor/positive_integer.h>
 #include <ydb/library/actors/core/actor_bootstrapped.h>
 #include <ydb/library/actors/core/log.h>
+#include <ydb/library/signals/object_counter.h>
 #include <ydb/library/signals/owner.h>
 
 #include <library/cpp/monlib/dynamic_counters/counters.h>
@@ -18,15 +20,16 @@ namespace NKikimr::NConveyorComposite {
 
 class TDequePriorityFIFO {
 private:
-    std::map<ui32, std::deque<TWorkerTask>> Tasks;
+    std::map<ui32, std::deque<TWorkerTaskPrepare>> Tasks;
     ui32 Size = 0;
 
 public:
-    void push(const TWorkerTask& task) {
-        Tasks[(ui32)task.GetTask()->GetPriority()].emplace_back(task);
+    void push(TWorkerTaskPrepare&& task) {
+        auto priority = (ui32)task.GetTask()->GetPriority();
+        Tasks[priority].emplace_back(std::move(task));
         ++Size;
     }
-    TWorkerTask pop() {
+    TWorkerTaskPrepare pop() {
         Y_ABORT_UNLESS(Size);
         auto result = std::move(Tasks.rbegin()->second.front());
         Tasks.rbegin()->second.pop_front();
@@ -63,15 +66,22 @@ public:
     }
 };
 
-class TProcess: public TMoveOnly {
+class TProcess: public TNonCopyable, public NColumnShard::TMonitoringObjectsCounter<TProcess> {
 private:
     YDB_READONLY(ui64, ProcessId, 0);
     YDB_READONLY_DEF(std::shared_ptr<TCPUUsage>, CPUUsage);
     YDB_ACCESSOR_DEF(TDequePriorityFIFO, Tasks);
+    YDB_READONLY_DEF(std::shared_ptr<TProcessScope>, Scope);
+
+    std::shared_ptr<TPositiveControlInteger> WaitingTasksCount;
     TAverageCalcer<TDuration> AverageTaskDuration;
     ui32 LinksCount = 0;
 
 public:
+    ~TProcess() {
+        WaitingTasksCount->Sub(Tasks.size());
+    }
+
     void DoQuant(const TMonotonic newStart) {
         CPUUsage->Cut(newStart);
     }
@@ -80,10 +90,15 @@ public:
         return Tasks.size();
     }
 
-    TWorkerTask ExtractTaskWithPrediction() {
+    ui32 GetTasksCount() const {
+        return Tasks.size();
+    }
+
+    TWorkerTask ExtractTaskWithPrediction(const std::shared_ptr<TWPCategorySignals>& signals) {
         auto result = Tasks.pop();
         CPUUsage->AddPredicted(result.GetPredictedDuration());
-        return result;
+        WaitingTasksCount->Dec();
+        return std::move(result).BuildTask(signals->GetTaskSignals(result.GetTask()->GetTaskClassIdentifier()));
     }
 
     void PutTaskResult(TWorkerTaskResult&& result) {
@@ -105,17 +120,20 @@ public:
         ++LinksCount;
     }
 
-    TProcess(const ui64 processId, const std::shared_ptr<TCPUUsage>& scopeUsage)
+    TProcess(
+        const ui64 processId, const std::shared_ptr<TProcessScope>& scope, const std::shared_ptr<TPositiveControlInteger>& waitingTasksCount)
         : ProcessId(processId)
-    {
-        CPUUsage = std::make_shared<TCPUUsage>(scopeUsage);
+        , Scope(scope)
+        , WaitingTasksCount(waitingTasksCount) {
+        AFL_VERIFY(WaitingTasksCount);
+        CPUUsage = std::make_shared<TCPUUsage>(Scope->GetCPUUsage());
         IncRegistration();
     }
 
-    void RegisterTask(const std::shared_ptr<ITask>& task, const TString& scopeId, const std::shared_ptr<TCategorySignals>& signals) {
-        TWorkerTask wTask(task, AverageTaskDuration.GetValue(), signals->GetCategory(), scopeId,
-            signals->GetTaskSignals(task->GetTaskClassIdentifier()), ProcessId);
+    void RegisterTask(const std::shared_ptr<ITask>& task, const ESpecialTaskCategory category) {
+        TWorkerTaskPrepare wTask(task, AverageTaskDuration.GetValue(), category, Scope, ProcessId);
         Tasks.push(std::move(wTask));
+        WaitingTasksCount->Inc();
     }
 };