+++ ydb/apps/ydb/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.22.1 ##
+
+* Fixed bug with state working in commnads `ydb workload query import` and `ydb workload clickbench import files`.
+* Fixed bug with parsing certificate files from profile.
 
 ## 2.22.0 ##